uploader_api/requirements.txt
FLASK==1.1.2
FLASK_LOGIN==0.5.0
FLASK_SQLALCHEMY==2.4.4
GUNICORN==20.0.4
PIKA==1.2.0
PSYCOPG2-BINARY==2.8.6
PYSMB==1.2.6
PYTHON-DOTENV==0.15.0
REQUESTS==2.25.1
WERKZEUG==1.0.1